inc: start to replace the infamous pfSense PHP module
Most of the stuff it does can be hand-rolled using ifconfig(8).
Since these operations are sparse and well-defined, there's no
harm in doing them.  The added benefit is a quantum leap in
transparency and traceability as well as using a script-based
language to do string manipulation for the command line as
opposed to C-style API fiddling (meaning: it's easier).