rc: add a safety sync after the recovery stage to avoid a vicious cycle
The current fix now works as expected and always brings back the
console login after an unclean shutdown, which was previously not
happening.  A small quirk, however, is that while the base system
recovers, the OPNsense backend on top may not fully recover directly
after the unclean shutdown.  In those instances, a manual reboot
will bring back the fully operational system and its services.

The unclean shutdown dilemma is, however, an unavoidable occurrence
and any kernel panic or power failure can corrupt the file system
into an unrecoverable state.  Having snapshots or backups at hand is
a good habit to get into, which can also 'fix' the eventual breakdown
of a physical disk.

Discussed with: ad@ jos@
Tested by: jos@