www: zap some spurious assigments
Most of the time, there is no return value from the functions,
sometimes $retval is not evaluated, and sometimes it just seems
totally convoluted to do it when 99% of the configuration code
does not do it and works fine wihout it.

It's also not practical, because there is no no notion of a
rollback that would have to ensue in such cases.

(cherry picked from commit 1eb5dae76d3ec1f78ac6a5c45e61f5caae37dc0c)