rc: persistent firmware cache dirs on MFS; closes #924
This adds the following directories to avoid problems with
smaller embedded systems that do not have enough RAM:

	/var/cache/opnsense-update
	/var/cache/pkg

This is particularly bad for Nano systems with enough RAM
as it may grind the flash storage, but since this is a
"luxury" problem with Nano the scope is limited to that.
It's either being able to upgrade the firmware for all or
not at all.

FWIW, firmware upgrades will have to move to persistent
storage as soon as we hit FreeBSD 11 for the simple fact
that we'll have to do partial upgrade stages for the ABI
jumps of the kernel / base system anyway and cannot expect
to fetch more packages after reboot.

Proper garbage collection exists in both tools that shall
avoid overloading the persistant storage.