cron: migrate the time-based rules cron job to autocron
o Gets rid of cron save/flush in the filter reload path.
o Gets rid of cron job injection into config.
o Move configure_cron() to the rules edit where it can be
  flipped in case of schedule changes.

To be totally frank, a 15 minute hook to deploy a time
based system is hardly accurate.  But what worries me more
is that we still have this patch in the kernel, totally
unsure of what it actually does when we have a schedule-based
reload on top anyway:

https://github.com/opnsense/src/commit/0a19f813177d18e569a457eff50a07c2e42abe04