dashboard: default to empty, allow empy save; closes #1048
We do have a config.xml default, and we don't need a second default
that pops up every time the old layout is cleared.  On top of that
the old layout could never be cleared again, so it only dealt with
an empty config.xml scenario which we don't have.