plugins: tweak interface plugging
PPPoE wasn't removed, which led to some refactoring to find
out why.  The outcome is the same, but works a little bit
different now.  While there, make it possible to omit the
config write in case we are going to call write_config()
shortly after anyway.

This will also be necessary for the legacy VPN plugins.