system: kill default gateway force override; closes #957
While we don't exactly know how this pans out, it's code that needs
to be inspected by forcing it to -devel for a while.  As far as I
can see, the following applies:

When there are no gateways, they are generated as defaults.

When they are edited, they stay in the config and their e.g.
choice of being disabled should be honoured.