system: really remove enableserial
We always set primaryconsole in our images, which is enough
for the code to figure out that serial was enabled.  If a
user changed this, he will have to reconfigure to get the
expected result: primaryconsole set to something different
other than serial and secondaryconsole not set will obviously
not be enogh of a hint for a system to use the serial one.