plugins: if a component has a faulty syntax...
simply eval() it and log the error instead of stopping execution.

We don't get to source the code, but since for strict plugins
there is no direct code execution the plugin acts as if it was
deinstalled.

As the plugin scope is growing rapidly, we need this to be able to
guarantee error-free execution in the widest possible sense.  This
doesn't currently protect function calls, but the same thing must
be done there as well.