plugins: lose the eval(), switch to include_once
Note that require_once is bad as well here, because it throws another
uncatched error as opposed to a warning when a file has not been found.
Now it is better in recovery, although 99% of the time this is supposed
to deal with syntax errors during bad core/plugin updates or manual
user edits.

It's also useful to note that the errors/warnings still make their way
to the crash reporter so we can do something about them.

Suggested by: @fabianfrz