mvc: try to fix crash report
Seems like years ago when PHP 5.5 had trouble iterating non-
existing arrays PHP 7 introduced a new problem with nodes
that seem to be set but no longer available, yielding:

    PHP Warning:  OPNsense\Base\FieldTypes\InterfaceField::eventPostLoading():
      Node no longer exists in /usr/local/opnsense/mvc/app/models/OPNsense/Base/FieldTypes/InterfaceField.php on line 132

The same thing happened in the acme-client plugin and was
allegedly fixed with the following code.  Even if this doesn't
help it is harmless and we'll see through the crash reporter
soon enough.