webgui: move settings around a bit #1347
Anti-lockout really has both SSH and web GUI as its targets,
which is a bit weird here.  Maybe we ought to split the options,
but for now move it a bit close to SSH.  A separate option makes
no sense at this point.  Maybe this is more of an advanced firewall
option?