Firewall, portforward. https://github.com/opnsense/core/issues/1326
Rename NatRule to ForwardRule, it seems better to split the logic for the different nat types and reuse the functionality in the base class. A portforward is actually a rdr entry combined with nat if relection is enabled, 1-on-1 nat is the exact opposite (nat, optionally combined with rdr). Which would make the parse logic to complicated to read if combined.