IPv6 Static over parent v4 PPPoE link
Certain ISP's allow static v6 address assignments even though dhcp6 works too. The only issue with using statics is that the setup for apinger gateway attempts to use the real interface, i.e. igb0, this does not work as the traffic is going via the pppoe link.

This patch addresses that and sets the gateway to correctly use the pppoe link.

This is an optional setting and an additional option has been added to the interfaces as a checkbox to ebable/disable this option.