Fixed DUID
Allow the user to fix the dhcp6 duid.

The existing DUID may be recalled and stored in the config file or a new DUID may be created.

On the call to function interface_dhcp6_configure the existince of a config variable 'system''global-v6duid' is checked, if it exists then the duid file is written before dhcp6c starts.

The loss of DUID on reboot if using RAM disk or the loss when upgrading or rebuilding is therefore mitigated.