dhcp6c - Use Reasons
Use the REASONS given by dhcp6c when it calls its script. This then only calls newipv6 when 'REQUEST' is the reason.

RENEW, REBIND or INFO do not have changes to the leases, therefore no call to update is needed and this prevents reloading which can affect VPNs, Unbound etc.

On RELEASE or EXIT, once again newipv6 is called as the lease addresses should have been removed by dhcp6c.