Set dhcp6c vlan Priority
Certain ISP's, Orange France for example use a VLAN, it is a requirement that the dhcpv6 packets are sent with a different priority, in Orange Frances case it is 6.

This PR allows the user to select whether or not to enable the setting of the VLAN priority for dhcpv6 packets and the value to use.

The is then used in the filter.lib.inc to add the extra set-prio value when the fillter is created.

This PR requires the the sysctl net.link.vlan.mtag_pcp be set to 1.