dhcp6 Client Adanced Options - Enter Prefix Interface Name
Currently, when using Advanced for dhcp6, the config file has the real WAN name as the interface on whuch the prefix will be applied, this needs to be a LAN ( real ) interface. Simply enter the interface name.

No error checking is carried out as this is already in an advanced selection.