dhcp6c handling
This commit corrects an issue when using advanced dhcp6c settings in that it now correctly configures the LAN interface(s) names. It also allows multiple LAN interfaces  to be used.

A complete break out of the configuration creation and script creation has been done, this is because to correctly allow for a change to a LAN interfaces that may affect tracking, the script creation MUST take place in the interface_configure routine  as a change to LAN tracking or not must be caught and the configs adjusted before calls to interfaces_dhcp6_configure.

New routines have been added that launch or reset dhcp6c plus kill and find & reset. These are all there to try and mitigate the chances of multiple dhcp6c clients. The kill_dhcp6c_client_process takes advantage of feature added to dhcp6c last year. Using the USR1 signal will force dhcp6c to exit cleanly but with no release. This is important, as a lot of the issues with dhcp6c are caused by timing. On a normal TERM  kill, dhcp6c will try to rleease any assigned values by signalling the server and wating for a response, this takes time. Thus if we are taking down the interface or making changes we want an immediate exit and as a new REQUEST will be sent we are not interested in releasing, thus dhcp6c exits quickly.

RTSOLD has been changed to a one shot. I am aware of a previous issue but the clean-up here around dhcp6c should mean that there is no requirement for it to be resident.

Updates include
sla id correction
Send ia-na0 mod
Send ia-na 0 was being sent even if prefix only was selected.
fix paths in rtsold script creation
Clean and polish
delete unused lock reference
fixes and changes rtsold to resident