pppoe Virtual IP fix #2
PPPoE VIP fix

First version broke IPv6, this corrects that and leaves things intact
unless the VIP is v4 AND is pppoe.

PPPoE VIP fix

With pppoe we do not know the gateway in advance, so wen ifconfig'ing, we determine if we have a pppoe connctiob on the interface and get the gateway from the gaeway file.