HTML Compliance - Attribute "content" value on Element <meta>
Error: A meta element with an http-equiv attribute whose value is X-UA-Compatible must have a content attribute with the value IE=edge.

The "Chrome=1" value was used to spawn Chrome Frame in IE 6,7,8,9.  However, "Google Chrome Frame was discontinued in January 2014", and think all IE<11 is beyond EOL.
So is this even needed/useful?

If needed/useful it would probably be better to support deployment by HTTP headers method instead.

Google Chrome Frame
https://en.wikipedia.org/wiki/Google_Chrome_Frame