MVC/templates, replace javascript_include_when_exists() for more universal theme_file_or_default() template function.
The legacy code will still use a plain file_exists() for the optional theme.js, which keeps both in a working state after this commit.
To be able to move more files into the template directories for https://github.com/opnsense/core/pull/2124 we need to keep existing plugins functional, so we could integrate these changes into a minor release.