dhcp6c & rtsold conf  & script creation breakout
This commit is a basic breakout of the conf and script file creation only. It does not fix the underlying issues with dupliate dhcp6c clients. It does fix an issue with advanced dhcp6c config and it also allows for changes in LAN config that may affect dhcp6c config.

Only call when required