interfaces: trust the GUI to do its job
There is still a weird interaction between this code and
get_real_interface as they both react to dhcp6usev4iface,
but now this is more obvious and could help with refactors
later on.

The upside is that we do not have to adjust this for another
IPv6 type that will gain IPv4 latch capability in the future.