firmware: first step to keep a list of known plugins #1663
While this works, even for retrospective reinstall, the real issue is
to get to know all the other plugins that are installed but not yet
known like an image-preinstalled os-dyndns.

Second step is how to present the plugins that may be missing in a
sensible way or clickable "install all missing plugins".