Auto Prefix Delegation tracking.
This adds the ability of prefix delegation to follow the change of the tracked interface.

Also added some changes to services_shcpv6.php in terms of help info and to limit the number of chars a user can enter when entering a value for PD ranges when tracking overide is in place. These changes reqiure a modified dhcp6c client that now gives out the prefix range and mask as an env var. This is picked up in interfaces.inc  and echoed to a file in /tmp.

Added correction for override ignoring other settings apart from range.