Add Active Schedule icon to firewall_rules display
The Firewall Schedules page shows a clock icon if the schedule is currently "active". This is great to see if your schedules are operating as you intended. This icon/feature is also great when placed on the firewall_rules.php page so you can tell if a firewall rule is active or not.

The html is taken directly from firewall_schedule.php so this new usage on the Firewall Rules page has the same look.

The "break;" statement is added so that $schedule is the actual schedule array member, for use in filter_get_time_based_rule_status().

I think the break should be added even without this icon display feature. I believe this foreach was unnecessarily iterating through the entire schedules array in all cases( even after it found the desired schedule and copied the description).