firmware: annotate HTTP(S) mirrors
Our mirror stays at HTTP for 18.7 as there will be issues
as reported in the forum already where HTTPS is "suddenly"
broken by SSL proxies in the way of the updates.

The final default HTTPS switch will happen with 19.1.  For
now use the development version and FreeBSD and derivates
still default to HTTP since the updates are signed.