make Schedule icon reflect current schedule state
On firewall_rules.php, there is no indication whether a schedule-based
rule is active. This change to the schedule icon applies the same styles
that are applied to a disabled/enabled Pass rule icon (text-muted and
text-success).

The break added to the foreach loop is needed to retain reference to the
attached schedule for the filter_get_time_based_rule_status() call