Add RFC3118 Option
This allows us to call functions defined in a new file. rfc3118_lib.inc which will generate the correct credentials where they need to be hashed or not. In the function in this PR, the function returns all of the request and send options required for  Orange France and fills in the client fields autonatically.

Also in the case of Orange France an extra entry box appears to allow the entry of the LiveBox_ID.

The PR is a work in progress, but all that is needed to nake it work for Orange France is the hash routine.

It;'s also possible to use the functions create the hex strings needed for other ISPs and auto fill the options as needed.