border between <a>´s are missing
I noticed that the separation of the individual <a> links is missing. Now it is up to you to say if we should install it again or not. if you do not want it please just delete this pr. have added two screenshots as a comparison. Thank you, best regards, René