firewall: repair log widget #2747
The initial design of the options could have been better.  Too
much smarts for little benefit.  Break "All" interface option
while here, we shouldn't use a text that could be an interface.

Also, remove the table widget layout fix since we have text-wrap.

But ironically enough Firefox doesn't care about text-wrap so
we are back to overflowing the container on long words.

It's still the same issue: too little space for displaying what
you want to display.  I'm not going to work on this anymore.