ui: another regression, change placeholder text to real text
In most cases the "placeholder" is a valid value so it doesn't
make sense to mark it as something less because that would mean
the user attention is drawn to it and more often than not these
unselected values are the good default behaviour so we might
draw users away from well-defined scenarios into support-heavy
edge cases.

Secondly, it looks more like disabled than placeholder.