resize function adapts to the sidebar height
since the navigation bar can change in height at any time due to the installation of plugins, and since we may want to redistribute the plugins, I have decided to dynamically adapt the resize function for navigation. now it does not matter how many menu items are in the sidebar. the resize function of sidebar navigation is now dynamic.