firmware: a bit backwards, not so bad actually
It's very difficule to improve UX of this page.  We now have a
separate changelog tab which helps with clarity over what is
"updated".

Now we can always go back to the last log of the past action.

And the log will not jump from the old content to the new one
when an action starts.

Two problems remain: the transition during "check for update"
looks quite empty.  And there is no user guidance as to where
he is and what he is will install in an overview shape.