unbound: add alias support for #3260
Break the rules of the Dnsmasq implementation while at it:

An alias can consist of a host and/or domain and/or description.
At least a host name or a domain need to be set, the rest will be
taken from the original entry.

Missing GUI parts in the override section...