system: disable warnings due to count() behaviour in PHP 7.2
We have been successfully using E_WARNING as a code quality measurement
since 2015 (see f6c6889d7).  The API components are even using E_ALL
with all its deprecated and strict checking so everything is an error.

There seems to be no change in count() other than throwing warnings now
which make it hard to write robust code as some parts seem to be buggy in
unreproducible ways.  The warning acts like a wrench in the gears so to
speak with no added code quality value.

Although count() bugs do exist in the code base we do not think that
PHP 7.2 should have created this warning under warning level since the
count() behaviour is relatively well-defined in undefined behaviour
cases and the strictness approach only leads to application damage with
no way but runtime to figure out it is or it is not working.

This is only for stable, not development.

Discussed with: @adschellevis