Routing, gateways. The `/tmp/*_defaultgw` construction has a race condition the way it is implemented now. for https://github.com/opnsense/core/issues/2279
It is used by dhcp client to detect if a default route might be overwritten and it determines default gateway priority. Since I don't want to refactor the dhclient-script at the moment, we best keep the file, but remove the "default" detection.
So system_default_route() sets the file, which dhclient can pickup when a new gateway is propagated.