add cache helper function in config.inc, needed for https://github.com/opnsense/core/issues/3567
get_cached_json_content() deserialize json data when the file is found and not yet expired, returns null otherwise.
Since there might be some spots in the code where this can be convenient, it seemed like a good idea to wrap it in a function within the legacy code base.
Maybe config.inc is not enterily the right spot for this, but util.inc is quite diverse already.