interfaces, cleanup. while working on https://github.com/opnsense/core/issues/3567
Since $realhwif is the parent interface, it doesn't make sense to check for _vlan. We might as well remove the check and leave the rest as is, since it only triggers when either media or mediaopt is set.

(cherry picked from commit 028e7a260da4cf1f93ef2d3f46715788e8cfc512)