Interfaces.php:  Remove visibility toggle on "Enable" control
COnfiguring an interface, or viewing its basic config,  just shouldn't be conditional on enabling it.

As the code stands, you can't even give the interface a name, unless it's also enabled. Much less set params that should apply to it. WHich often you may want to set, even if the interface isn't yet ready to be enabled.

Basics like this should always be accessible for editing, there's no reason they should be behind a visibility toggle and blocked unless enabled.  It's not like the form is long or complex either.  This PR removes the unhelpful show/hide behaviour.