Document "type" field which is omitted in the list. Add type = "section_title" as an option
The type field is omitted in the list of valid fields.

Also there is no way currently to add "section headers" to groups of fields. While this doesn't have syntactic importance, it is useful for UI purposes and sections are widely used elsewhere in the UI *(eg: `System: Settings: Administration` to add sections headers for Web GUI, Secure Shell, Console and Authentication)*.

This also ensures that Phalcon generated pages have consistent appearances for section_titles in the UI, and can easily be added if required.