Allow for variable debug info from dhcp6c. None, Level 1 or Level 2
In the current code, debug is either -d or -D, these are both debug levels, Not settings a debug flag at all results in zero debug output. Not totally happy with the way I've implimented the  command string in interfaces.inc, but it works, I could not get it to work if the debug was nothing and part of the array. If someone has a better way, go do it.