dhcp6c multiwan support
Allow multiwan support for dhcp6c. All interfaces must be scanned to see whether dhcp6c is in use on that interface, the conf file is appended on each iteration. I have allowed for advanced mode too. In  file override mode there can be only one file that must handle all interfaces, thus I need to add a check that if an overide file is specfied on any interface, all other interfaces must show that the override is in place.. WIP.

However, the overide may no longer be needed as you can do everything without the need to overide the file. Tested and working on my system. Work may be needed around the firewall, but on my test system it all apears to work fine.