Added extra checks around the start or sighup of dhcp6c
Still occasionally seeing multiple dhcp6c instances, this has to be made impossible. Now checking to see if the pid matches the process and if not then kill the process and delete the pid ( if it exists ).
Change text so it says starting dhcp6c client. The client runs on all interfaces, there is ONLY one.