Further tighten up creation of dhcp6c.conf. Create entry ONLY if the interface is up.
When NEWWANIPV6 gets called on an interface change, the existing code will create a config for dhcp6c even if that interface is down, checks are added to see if its down and config for that interface is skipped if thats the case. DHCP6C removes addresses from the interfaces when it exits unless NORELEASE is set.