re-add 'domain' to dhcpdv6 interface configuration
'domain' was removed in #3824 because it is not used by dhcpdv6. But it is used by unbound and dnsmasq for DNS registration of DHCP static mappings. Just set it to the first entry of the domain search list.

(cherry picked from commit 333c2eb63d185c5eab7cbf5e9882fff91d733999)