Reporting_Health_Resolution_Description_-_Units_Singular
Change: (Current detail is showing 10 minutes averages.)
To: (Current detail is showing 10 minute averages.)

Plural vs Singular Examples:
The mini-series will be aired as ten segments 2 hours each.
The mini-series will be aired as ten 2 hour segments.

There is a wait time of 72 hours.
There is a 72 hour wait time.

The cancellation policy is 3 days.
There is a 3 day cancellation policy.

This is a 3 year project.
The project duration is 3 years.