Select better binding address
Check the interface IPs to preference a binding address that is in the same subnet as the gateway monitor address, rather than picking the first address. This fixes an issue for when CARP is used with a single WAN IP address and the adapter's first address is a private one.