firmware: ignore man page database regeneration
Looks like on the running system and in the build system the values
are static but obviously going from build to running system the
regenerate causes the checksums to shift.  Not a security issue for
the "man" page databadse so better to hide these files from the
audit to avoid confusion and questions.

PR: https://forum.opnsense.org/index.php?topic=18484.0