interfaces: l2tp/pptp cleanup; closes #2707
The interfaces are created by mpd5 daemon during configuration.
We do not seem to have more than one issue here: IPv6 disable
is too soon but that is easily fixed.  Device creation moves
on ok and there is no apparent error in functionality with or
without it.

While here fix the link when a $pppid is set and properly protect
the redirect url.  Previously empty() was too strict and it ignored
$pppid of zero.

Also remove spurious "IP Address" help text from PPP device
configuration.