Firewall: add model for categories. work in progress for https://github.com/opnsense/core/issues/4587
refactor javascript code and move it into opnsense_legacy.js for reuse. while here add a standard controller we can use to query existing categories and allow usage from the components in scope of this change.
Keep current selection in the user session so selections are sticky among tab changes and edits. The previous "hack menu" was a bit impractical to expand to other components as well.