firmware: do not allow plugin install if remote repo version differs #4500
We need the matching mirror version for the plugins to install so
simply block the update and let the user update first (instead of
only checking for updates and then installing later versions of
plugins).

This was a larger problem in past years but it is good practice to
require an up-to-date system anyway.