Fix Icon Toggle for Block & Reject
When a rule is set to Allow, Block or Reject, the icon is `fa-play`, `fa-times` and `fa-times-circle` respectively.  When you toggle rule to disable it, this code changes the icon to `fa-play` (Allow Icon) regardless of it being Block or Reject, which is confusing.  Since `fa-spinner fa-pulse` is appended to the class, it takes priority for the icon.  When removed, it restores the proper icon with the appropriate `text-muted` or `text-success`.  There is no reason to remove and add `fa-play` - it only breaks the icon toggle.