Prefix delegation fix for dhcpdv6
Currently any PD delegation is incorrectly using the same prefix as the parent interface. For the avergage user who does not need PD on thier LAN this is not an issue, it  will cause issues for those who do want to use dhcpdv6 and give out prefixes. It is quite complex to do as the PDs will vary depending on the prefix obtained from  the ISP and the delegation size being given out, not a simple 'just add '1;
Added multi wan capability
Correct Manual override issue