IXR_Library: do not trim xml values, there's no reason for. closes https://github.com/opnsense/core/pull/4707
xml-rpc isn't used very often these days, it proably doesn't make sense to try to upstream this change since the library doesn't seem to be very active anymore.
When values are of non string type, it makes sense to trim() for safety, for strings leading spaces can have a meaning.