VPN / IPSec: calculate netmask for provided tunnel addresses when using VTI.
For IPv4 this should be backwards compatible with the previous code, since the netmask isn't really used to determine if the other end is reachable (it seems so at least), for ipv6 some consumers a valid netmask
seems to be required in order to function properly (frr). Since ipv6 doesn't seem to support setting a netmask in combination with a destination address and the other end apparantly doesn't really care, we just set an address with a mask in case it's an ipv6 tunnel.

for https://github.com/opnsense/core/issues/4700