webgui: clarify help text for firewall rules traffic direction
A lot of OPNsense users become confused by the concept of traffic direction in firewall rules (despite this being covered in the official documentation). This change is to expand the help text that is shown in the webgui for the Direction option in the firewall rules, along the lines of the official documentation, to hopefully to provide more clarity for users. Feel free to accept or decline in your discretion!