Update README with further module order info
This additional info is drawn from the lighttpd module docs. 

Given webgui.inc already loads some modules by default, I can see a potential situation where a particular module is not able to be loaded through the include directory in the desired order relative to the default modules. Not sure there is a neat solution to that.