Traffic: Fix long comment preventing ipfw reload (#5023)
It seems that comments are limited in ipfw to 80 chars which is not
documented (at least not in the manpage `ipfw(8)`). If users
accidentally use long traffic shaper rule descriptions this limit could
be reached causing updates via the GUI apply button to be silently
ignored which will cause confusion.

`ipfw /usr/local/etc/ipfw.rules` shows the following error:

    Line 77: comment too long (max 80 chars)

and exists with edit code 65.

Interestingly, a slice of up to `0:83` still works. As this is
undocumented and the error says max 80 chars I go with something
slightly below this max to account for the whitespace.