Services / DHCPvX - refactor dhcpd_staticmap() so it takes empty (descriptive) only leases and protocol family into account.  closes https://github.com/opnsense/core/issues/5264
Although dhcpd_staticmap() is a great idea from the resolvers point of view, the status pages have a bit different requirements. The easiest fix is to merge requirements, although this has the downside of making the function a bit more convoluted, which is a price we probably better pay for now.

My original suspicion that the line "if (!empty($value) || $key == 'start' || $key == 'end') {" looked odd seemd to be right, as it will omit lease information when actually there.