firmware: deleting the file under configd execution and copy...; closes #5396
... leaves us with permission 640 even though we have copied a
644 file.  Removing the unlink() makes this work without a
chmod but the unlink is there for the fact that /etc/ssl/cert.pem
used to be a symlink and could clobber the actual file linked
which was the original package provided.

Might be an umask issue, but better leave it where it is.

(cherry picked from commit 7a68bab085944f98992b523ae932e0e4119476ee)