Initial Commit
Option to use ULA VIP in dhcpdv6 server when using tracked interface. dhdpdv6 override must also be enabled. VIP is auto detected and option is only visible if a VIP address is enabled on the interface. Config files are written for dhcpdv6 and radvd. Will probably require some sanity checks.