system: display system tunables as well
This sort of completes the previous effort to improve the
sysctl handling.  The idea to embed sysctls into the config.xml
is certainly valid, but for some values we needed more leeway
for system defaults and changing them over time, which lead
to this state where we can drop config.xml default tunables
since the system will now apply and display then for the same
visibility but offer less incentive to change these values.

One thing missing here is adding a system tunable override
(plus sign) for system values, but we shall deal with this later
because for now they can always be added manually by copy+paste.