dhcp: kill "static" mode, move flags to advanced #5185
While technically not correct to make flags "AdvDeprecatePrefix"
and "AdvRemoveRoute" we simply use it to overwrite the existing
defaults which depend on CARP-or-not situation in the address
being used.

There was an extensive discussion about what it should and should
not do but the only way to keep the current defaults AND provide
requested overrides is this one.

The valid values are either "off" or "on", but no validation
takes place as customary to this page's advanced options. Use
with care.