console: a small audit and LAGG functionality test
Seems to work fine, but the new and old code has some oddities like
obsessing over the "up" flag, but we set all interfaces up before
we start the process.  Some style updates and separation of use
while here.

Splitting on /\d/ seems overly toxic when the device is called
e.g. em10_vlan12 so address that in the utility code where it is
carried out.  get_interface_list() is probably a function that
should be removed in the mid-term.