interfaces: loopback "lo0" exists for VIPs
Historic oversight of wedging in a loopback interface in virtual IP
section while not making sure the actual configuration receives a
virtual IP when it needs to.  This was previously fixed by a batch
VIP regeneration in interfaces_configure() but we already saw that
this is not needed in 99% of the cases.

PR: https://forum.opnsense.org/index.php?topic=27547.0
(cherry picked from commit 9a618ba6893a7a8b457439609a15df40fb360b68)