interfaces: suspicious code is suspicious
I'm not sure why interface_bring_down() is needed when both
IPv4 and IPV6 are empty.  It sort of means "handle this the
hard way when doing tunnel configurations" althoug the code
disagrees about the historic comment and the code that was
introduced... "set to none" vs. "ipaddr <> none" and later
"empty(ipaddr)" to match the comment.  In the grand scheme of
things this does not matter at all...

So in 22.1.3 we removed the inline configuration of GIF and GRE
which causes interface configuration to strip the addresses
added by device configuration instead of refixing it on the
fly (executing code twice all the time).  The code flow was
always correct but in practice tripping over itself so now try
a more sensible approach by stripping addresses when we have
assignments going on individually for IPv4 and IPv6.

PR: https://forum.opnsense.org/index.php?topic=27553.0
(cherry picked from commit 61500f6790969ba1c426c27322ba8879dd96bd5a)