system: dpinger support for IPv6 aliases #5777
Since gateways support VIPs we need dpinger to support it too.

The implementation is a little different asking for runtime address
of the interface and trying to match a subnet before falling back
to the standard address.

IPv4 slightly adjusted to follow the same pattern, but without the
pin to an explict VIP address.