openvpn: move device handling to plugins facility #5411
While here remove a couple of references from loops and
improve coding style while at it.

While openvpn_prepare() is not the same as reconfiguring
a device (pluginctl -d ovpns1 vs. pluginctl -s openvpn 1)
we can get away with it as used before.  It might be the
subject of later work but for the sake of clearing the
scope of interfaces_configure() this is more than enough.