system: reduce $pwread scope to stop PHP from being silly
On failure of "pw" command the explode would split
"pw: no such user `foobar'" and then later check it.
If the user name was "pw" that would be suboptimal
for eligbility.

Instead ignore stderr and do not fill $userattrs which
is already handled by the if below but make it more
explicit using === null comparision.