interfaces: upon reload request also generate the $realhwif; closes #5819
This goes for all "known" devices.  We need to be careful of reload loops
that create a device and then reload it which could trigger another interface
configure, but at the moment at least all $reload consumers are outside of
the plugin device code / single action driggers.

Also see: 90f471cb19