interfaces: simplify the staticarp code and make the problem worse #5637
According to dhcp server the static arp setting persists through a
disabled server but doesn't trigger on disabled or when the staticarp
flag is actually set.  To keep the current flow make the configuration
unconditional which makes the problem we are trying to solve worse.

There are 3 callers to the function interfaces_staticarp_configure()
and rc.linkup is one that is not called when booting so it's not our
problem.  The remaining two in interfaces.inc and dhcpd.inc should get
to see if staticarp needs to be flipped and thus execute the arp flush
as well under this particular condition.