VPN / IPsec - fix cleanup regression in https://github.com/opnsense/core/issues/4460
As we stopped using "required" in our spd entries we need other means to remove previously manually added ones.
This commit collects all policies that are likely inserted manually and removes the ones that are being used in active phase 2 entries (reqid) configured with manual entries.

Combined with the new diagnostics page a user should be able to manually remove entries we couldn't automatically cleanup due to the risk of removing unrelated manual entries.

Also cleanup the logging a bit as the previous messaged where added for temporary use.