interfaces: same bug as previous
It would be better if we didn't clean up as a side effect when running
a per-device configuration but the way vxlan/loopback were written that
only happens in batch mode so at least try to keep the other devices
as is and clean up any stray objects.

We could push this further, but as we can see the risk of regression
is real so do not try to touch the code any further for now.

(cherry picked from commit ddb4af9040d430889c5b365cbce2c7f93eada872)