interfaces: disable reload with same IPv6
rtsold resolvconf handing forces a lot of reloads now, something
also seen in rc.newwanip -- let's go the extra mile and lock the
cache IP in place until we do a full reconfigure.

We may have to build some sort of "expire" feature for the cached
IP since between forced reconnects we will want to reload again
anyway.  But not sure where that is as rc.linkup is unreliable as
it has been messing with this before.