firewall: extend pool options support (https://github.com/opnsense/core/issues/5934)
Even though we're currently not able to offer other options than round-robin as pfctl seems to prevent that, it might still be a good idea to make the current options explicit and leave room for future enhancements.

This commits adds poolopts to gateway groups and explains options. If users want to mix sticky and non-sticky, this will be possible with this change. Other options (random, source-hash, ..) unfortunately not.