interfaces: remove is_interface_wireless() #5987
Make the presence of <wireless/> node authoritative except
for the assignment page where we need to set this node in
the first place.

Now pivot away from a handrolled regex of devices names to
trust the output of the sysctl net.wlan.devices which is also
the prefix for our clones.