mvc: translate base field error
Without gettext() wrapping we never produce a string in the
translation for it.  It gets passed to gettext() twice then
but that is matters not as the target string cannot be translated
again when called a second time.

(cherry picked from commit edb286cf6a9dd08ab1a549b2d06fdc845d341838)