firmware: in retrospect always fetch the signature file
When we move to the next major the file contents for
the txz are all the same but their signature is not.
So until the next changelog sync was carried out the
CHECKSUM matched and the signature will stay stale.

(cherry picked from commit 607faca9bab3f1a1cee956c4e6537a92716e8dc6)