interfaces: for now bring back duality of PPPoEv6 and other types
This doesn't invalidate the PPPoEv6 mode, but it does keep the race
between router files for different types.  In this case e.g. Telekom
ISP will require the IPv6 from PPPoE to be set in order to be able
to do DHCPv6 for whatever reason.

PR: https://forum.opnsense.org/index.php?topic=31955.0