system: let our functions deal with null == '' as they used to
There is no point in adding control flow to callers that lead
to the same result in edge cases anyway.

Some parts of the code seem to be abusing ip_in_subnet() but
we can just return false in that case as the IP is not in the
empty subnet.