Unbound: small optimizations
No need to register callbacks when stats are not enabled and no need to use
a circular buffer on the logger side since we don't need to append left. This avoids
the need for converting it to a list.

(cherry picked from commit 3170c77880a5673f0444dc31ccda67ba95e37103)