Unbound: simplify policy matching
Since we don't need to account for whitelisted domains in the dnsbl_module anymore,
both the configuration and the algorithm for policy matching can be simplified significantly.

In this case the assumption will always be that if there is a "bypass" in the dnsbl, those
IPs will get excluded from blocking, the downside of this strategy is that we cannot
allow those passed queries to get into the cache, else subsequent requests from IPs that
should be blocked will still pass.