unbound: dnsbl_module: prevent logging SERVFAIL twice
in cases where e.g. an internet connection is down, a reply will not be present should unbounds' iterator module return.
normally we marked this as a SERVFAIL, but Unbound already does this for us in the servfail callback.
This means entries were logged twice, once with a "Pass, Recursion, servfail", another one with "Drop, Local, servfail".
This is ambiguous and would skew the relevant statistics.