interfaces: also set PCP value on IPv4 DHCP traffic
This requires to set up the VLAN priority field of the
DHCPv4 settings as well as removing the vlan-pcp option
which is redundant now.  This way we have a simpler
approach to finding the correct value while not borrowing
from IPv6 or parsing the advanced options for vlan-pcp.

Leaving the old vlan-pcp in place should not matter.
The last value is ours and this one should be used.

PR: https://forum.opnsense.org/index.php?topic=33376.0
(cherry picked from commit d195cd2e8a315ccd2069c01507cbf258793630f7)