System: Gateways: Group - POC support for "gateway_interface" option on groups (https://github.com/opnsense/core/issues/6486)
At the moment this doesn't look like a viable option as it's too easy to invalidate the ruleset leading to loading errors like "all pool addresses must be in the same address family". Technically it's not very difficult to push the correct options, but as <interface> <address> could be combined with <interface> alone, the validation seems to be rather complicated. Forcing the items of a group (addresses only or interfaces only) might solve some of the issues, but is quite difficult to implement knowing the different parts in play here.