interfaces: RENEW and REBIND report partial PDINFO #6583
Previously it was thought that RENEW/REBIND do not reload properly but
this was eventually about another bug for when the server returns "no
binding" message and subsequently fixed in dhcp6c but not yet released
up to 23.1.8.

Ignore PD RENEW/REBIND as it will be the same value as REQUEST anyway.