src: legacy_bindings.inc already included by config.inc
client_connect.php also includes it but not config.inc.  Try to leave
it at that to not pollute it unnecessarily.  The other scripts might
be able to do it as well this way but don't fix something that is not
broken either.

(cherry picked from commit b7db4316fd380e3da1f5f761cfe2fefa5aefffa3)