dhcp: refactor LeasesController
hopefully the last bit, this will drop parsing the config
for configured dhcp interfaces and will instead match
it on the output of ifconfig. While this would normally not
cause issues for ipv4, this is mostly done to be symmetrical
with the upcoming ipv6 leases change, where it is much more
likely for interfaces to be missed.