dhcpd6: refactor LeasesController
like the ipv4 refactor, mapping an IP to an interface
is now done using ifconfig. For ipv6, the chances of an
interface being missed using only NDP and configuration
data are pretty high. This is largely due to most
setups not having a static IP configured, but are instead
set to track an interface with DHCP functionality enabled
under the hood by default. NDP data is also ephemeral by
nature