Fix IPsec key pair generator for secp256k1 EC and add properer naming to UI (#6817)
Key generation for the "256" EC does not work as secp256r1 does not
exist in OpenSSL, it's called secp256k1.

Also the names shown in the UI are ambiguous as there are several curves
with 256, 384 or 521 bits.

(cherry picked from commit eafbf81b248e694641df5c2b3293b9a55457b272)